requirements.txt
streamlit
bcrypt
fpdf
openai
python-dotenv
plotly
PyMuPDF
groq
reportlab